Add svg mods and move SVG-related types into them.
